TST: borealis general classes
Consolidated Borealis test classes:
- placed in their own utility file,
- removed duplicate tests using subTest, and
- ensure all version/structure combos are run using test classes.
